Refactored hook component and split it into multiple files. Still a bit of mess
